requirements.txt
requests
python-telegram-bot==13.13
python-dotenv